requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
starlette>=0.36.0
httpx>=0.27.0
click>=8.1.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
anthropic>=0.40.0
sse-starlette>=2.0.0
